requirements.txt
Flask~=1.1
markupsafe==2.0.1
pyrebase4==4.5.0
Flask-FontAwesome==0.1.5
requests==2.28.0
docker==5.0.3
#gunicorn